requirements.txt
ultralytics>=8.3.0
torch
torchvision
numpy
pandas
opencv-python
pyyaml
rich
wandb
